+++ trading_bot/websocket/__init__.py
@@ -1,4 +1,5 @@
 from trading_bot.websocket.client import BinanceWebSocketClient
+from trading_bot.websocket.kline_aggregator import BinanceKlineAggregator
 from trading_bot.websocket.price_cache import PriceCache, get_price_cache
 
-__all__ = ["BinanceWebSocketClient", "PriceCache", "get_price_cache"]
+__all__ = ["BinanceKlineAggregator", "BinanceWebSocketClient", "PriceCache", "get_price_cache"]